requirements.txt
# Binance API
python-binance>=1.0.19

# Data Analysis
pandas>=2.0.0
numpy>=1.24.0

# Technical Analysis
ta>=0.11.0

# Environment Variables
python-dotenv>=1.0.0

# Colored Terminal Output
colorama>=0.4.6

# HTTP Requests (dependency of python-binance)
requests>=2.31.0

# Web Dashboard
flask>=3.0.0
flask-socketio>=5.3.0
eventlet>=0.33.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
joblib>=1.3.0
# Optional: tensorflow>=2.15.0  # For LSTM (uncomment to enable)
